requirements.txt
torch
opencv-python
mediapipe
numpy
matplotlib
scipy
torchvision
Pillow
requests